+++ src/ZoneTree/Segments/Disk/DiskSegmentCreator.cs
@@ -6,7 +6,7 @@ namespace Tenray.ZoneTree.Segments.Disk;
 
 public sealed class DiskSegmentCreator<TKey, TValue> : IDiskSegmentCreator<TKey, TValue>
 {
-    readonly int SegmentId;
+    readonly long SegmentId;
 
     readonly ISerializer<TKey> KeySerializer;
 
@@ -28,7 +28,7 @@ public sealed class DiskSegmentCreator<TKey, TValue> : IDiskSegmentCreator<TKey,
 
     public bool CanSkipCurrentSector => false;
 
-    public HashSet<int> AppendedSectorSegmentIds { get; } = new();
+    public HashSet<long> AppendedSectorSegmentIds { get; } = new();
 
     public DiskSegmentCreator(
         ZoneTreeOptions<TKey, TValue> options,
