+++ src/ZoneTree/Segments/Disk/MultiSectorDiskSegmentCreator.cs
@@ -6,7 +6,7 @@ namespace Tenray.ZoneTree.Segments.Disk;
 
 public sealed class MultiSectorDiskSegmentCreator<TKey, TValue> : IDiskSegmentCreator<TKey, TValue>
 {
-    readonly int SegmentId;
+    readonly long SegmentId;
 
     readonly ISerializer<TKey> KeySerializer;
 
@@ -34,7 +34,7 @@ public sealed class MultiSectorDiskSegmentCreator<TKey, TValue> : IDiskSegmentCr
     
     TValue LastAppendedValue;
 
-    public HashSet<int> AppendedSectorSegmentIds { get; } = new();
+    public HashSet<long> AppendedSectorSegmentIds { get; } = new();
 
     public int CurrentSectorLength => NextCreator.Length;
